Created some plots. Check description for notes
The last, graph I was trying to create is for rent vs airbnb listings by zipcode year over year. But it's way too loud at the moment so I still need to find a way to rework, probably have every year as it's own plot? and also show year prior to 2015